requirements.txt
# Flask and related extensions
Flask==3.0.2
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Firebase integration
firebase-admin==6.4.0

# Google AI (Gemini)
google-generativeai==0.3.2

# Async support
aiohttp==3.9.3
asyncio==3.4.3

# Environment variables
python-dotenv==1.0.1

# Date/Time handling
pytz==2024.1

# Security
cryptography==42.0.5

# Utils
requests==2.31.0
validators==0.22.0

# Database
SQLAlchemy==2.0.27  # For local caching if needed

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5

# Development tools
black==24.2.0  # For code formatting
flake8==7.0.0  # For linting

# Additional dependencies
Authlib==1.3.0
Adafruit_DHT==1.4.0  # For DHT sensor support
RPi.GPIO==0.7.1  # For Raspberry Pi GPIO support
